fixed cython compile error and np.int

+++ src/pyuoi/mpi_utils.py
@@ -8,7 +8,7 @@ try:
     from mpi4py import MPI
     _np2mpi = {np.dtype(np.float32): MPI.FLOAT,
                np.dtype(np.float64): MPI.DOUBLE,
-               np.dtype(np.int): MPI.LONG,
+               np.dtype(int): MPI.LONG,
                np.dtype(np.intc): MPI.INT}
 except ImportError:
     pass
